requirements.txt
mcp
httpx
python-dotenv
psutil
pywifi
comtypes
python-whois
uiautomation
opencv-python
pyautogui
Pillow
stegano
pyttsx3
